+++ implementation/js/src/infrastructure/lib+-/marionette/layout.js
@@ -187,7 +187,7 @@
 						}
 
 						//Reusable view?
-						var Reusable = app.get(name, _.isPlainObject(options)?'Widget':'');
+						var Reusable = app.get(name, _.isPlainObject(options)?'Widget':'', true);
 						if(Reusable){
 							//Caveat: don't forget to pick up overridable func & properties from options in your Widget.
 							this.show(new Reusable(options));